feat: only restore router cache once per popup session
The purpose of the router cache is to restore the last visited route
during the startup of the popup, but without any explicit check
the cache would try to restore the route an inifinite number of times.
Because the router cache is never used to restore a route any other time
it is safe to assume that we only want to run the restore function once.