Use playId to track all creations within a worker.
This simplifies scenes and allows tracking of entities created during tests as well as those set up for tests.

However, we need to override the browser's `fetch` method to accomplish this. I have added a page extension that does this, but it is possible to create pages from browser contexts, contexts from browsers and even browsers from browser names. We simply need to expect imperfect coverage on this header.