fix(browser): use capture for click listeners
On medium.com they are using 2 event listeners on  events that use the  property.  This makes it so clicking on  on the autofill overlay not open the pop up.  This works fine on other sites.  This change fixes the issue so medium.com is not overriding us opening the pop up by using  on our click listeners.